feat: extract runMigrations to its own module and add replay tests
Extracting runMigrations + NUMBERED_MIGRATIONS + runNumberedMigrations
from db/index.ts into a new db/migrations.ts lets tests import the
migration runner without dragging Electron through data-dir. The
logger already handles non-Electron contexts via lazy createRequire.

Adds tests/migrations/replay.spec.ts with four assertions:

1. **symmetry**: SCHEMA + runMigrations ends up with every expected
   table (including llm_calls, schema_version) and every column the
   legacy ALTER block ensures. Catches "added column to SCHEMA but
   forgot the matching migration" or vice versa.

2. **idempotency**: a second runMigrations call is a no-op — same
   versions applied, same column set.

3. **replay**: starting from a "pre-numbered-system" DB (SCHEMA
   applied, llm_calls and schema_version dropped, accounts row
   present), runMigrations bootstraps the numbered system, applies
   every migration, and preserves the seeded data.

4. **sequential versions**: NUMBERED_MIGRATIONS versions are
   contiguous starting at 1.

New Playwright project `migrations` in playwright.config.ts. Run with
`npx playwright test --project=migrations`. All 1382 existing unit
tests still pass.
